requirements.txt
# FaxTriage AI Backend Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
anthropic>=0.18.0
PyMuPDF>=1.23.0
Pillow>=10.0.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
